Add initial structure and 2 commands for stateful testing
This commit adds the stateful testing harness structure and two commands. It
adds the signer test initialization to the test context constructor, and defines
the BootToNakamotoCommand and SkipCommitOpSecondaryMinerCommand, aiming to split
the `allow_reorg_within_first_proposal_burn_block_timing_secs` test into atomic
operations. As described in 1cb033c, the end goal of this effort is the ability
to run all the atomic commands in either seed-based random sequences, or in
a predefined order. The `hardcoded_integration_test_using_commands` showcases
the previously mentioned integration test's translation to the command-based
stateful testing approach.